Move route for new donation to a controller class
